requirements.txt
streamlit
pandas
pillow
google-auth-oauthlib
google-auth-httplib2
